Create README.md
I’m a psychology graduate with a passion for WordPress design and development. I enjoy UX/UI design and am currently learning the basics of HTML and CSS. I'm also expanding my skills in Figma and Azure. My goal is to become a business analyst, and I’m actively learning about the digital landscape and how to excel in this ever-evolving field.